src/requirements.txt
fastapi>=0.118.3
uvicorn>=0.38.0
mlx-whisper>=0.1.0
python-multipart>=0.0.6
Jinja2>=3.1.2
ffmpeg-python>=0.2.0
psutil>=5.9.0
python-dotenv>=1.0.0